update cstm formation to ensure it is float and not int

+++ pyyeti/nastran/op2.py
@@ -3250,7 +3250,8 @@ def rdpostop2(
                         [
                             [+0, 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 0, 1],
                             [-1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                        ]
+                        ],
+                        dtype=float,
                     )
                     cstm = np.vstack((bc, cstm))
                     continue
